% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ann}
\alias{ann}
\title{Cell type annotations for Pollen data}
\format{
A character vector with cell type labels:
\describe{
\item{length}{Number of cells in the dataset}
\item{values}{Cell type names as character strings}
}
}
\source{
Extracted from \code{colData(PollenGliaData())[["Inferred Cell Type"]]}
}
\usage{
ann
}
\description{
Character vector containing inferred cell type labels for each cell
in the Pollen dataset.
}
\examples{
data(ann)
table(ann)
length(ann)
}
\keyword{datasets}
