% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateScEnsemble.R
\name{CreateScEnsemble}
\alias{CreateScEnsemble}
\title{Create a ScEnsemble object}
\usage{
CreateScEnsemble(sce, annotation = NULL)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{annotation}{A character vector of true labels (optional)}
}
\value{
A ScEnsemble S4 object
}
\description{
Initializes a ScEnsemble object from a SingleCellExperiment and optional annotation vector.
}
