% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onAttach}
\alias{.onAttach}
\title{Package startup function}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Character string giving the library directory}

\item{pkgname}{Character string giving the package name}
}
\value{
Invisible NULL
}
\description{
This function is called when the package is loaded and provides information
about optional dependencies that may not be installed.
}
\keyword{internal}
