% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EnsembleResults-class}
\alias{EnsembleResults-class}
\title{EnsembleResults Class}
\description{
An S4 class to store ensemble clustering results
}
\section{Slots}{

\describe{
\item{\code{ensemble_clusters}}{A list of ensemble clustering results}

\item{\code{ensemble_ari}}{A list of ensemble ARI scores}

\item{\code{ensemble_quality}}{A list of ensemble quality metrics}
}}

