% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IndividualResults-class}
\alias{IndividualResults-class}
\title{IndividualResults Class}
\description{
An S4 class to store individual clustering results
}
\section{Slots}{

\describe{
\item{\code{clustering_results}}{A list of clustering outputs}

\item{\code{ari_scores}}{A list of ARI scores}

\item{\code{runtimes}}{A list of runtime measurements}

\item{\code{errors}}{A list of error messages}

\item{\code{embedding_data}}{A list of embedding data}
}}

