% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pollen}
\alias{Pollen}
\title{Pollen single-cell RNA-seq data}
\format{
A SummarizedExperiment object with:
\describe{
\item{assays}{Gene expression count matrix}
\item{colData}{Cell metadata including cell type annotations}
\item{rowData}{Gene metadata}
}
}
\source{
\code{scRNAseq::PollenGliaData()}
}
\usage{
Pollen
}
\description{
Single-cell RNA-seq data from pollen grains, obtained from the scRNAseq package.
This dataset contains expression data for multiple cell types in pollen development.
}
\examples{
data(Pollen)
dim(Pollen)
table(colData(Pollen)$"Inferred Cell Type")
}
\references{
Li et al. (2017). Single-cell transcriptomes reveal characteristic features
of human pancreatic islet cell types. EMBO Reports, 18(9), 1589-1600.
}
\keyword{datasets}
