% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ScEnsemble-class}
\alias{ScEnsemble-class}
\title{ScEnsemble Class}
\description{
An S4 class to store data and results for ensemble clustering
}
\section{Slots}{

\describe{
\item{\code{sce}}{A SingleCellExperiment object}

\item{\code{annotation}}{A character vector of true labels}

\item{\code{individual_results}}{An IndividualResults object}

\item{\code{validation_metrics}}{A ValidationResults object}

\item{\code{hypergraphs}}{A list of hypergraph representations}

\item{\code{ensemble_results}}{An EnsembleResults object}
}}

