% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ValidationResults-class}
\alias{ValidationResults-class}
\title{ValidationResults Class}
\description{
An S4 class to store validation results
}
\section{Slots}{

\describe{
\item{\code{validation_indices}}{A list of validation indices}

\item{\code{normalized_indices}}{A list of normalized indices}
}}

