% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{calculate_all_validation_indices}
\alias{calculate_all_validation_indices}
\title{Calculate Validation Indices}
\usage{
calculate_all_validation_indices(object, ...)
}
\arguments{
\item{object}{ScEnsemble object with completed individual algorithms}

\item{...}{Additional arguments}

\item{verbose}{Logical, whether to print progress messages}
}
\value{
ScEnsemble object with validation_indices slot filled
}
\description{
Calculate validation indices for clustering results
}
