% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ensemble_clustering}
\alias{ensemble_clustering}
\title{Ensemble Clustering Algorithms}
\usage{
ensemble_clustering(object, ...)
}
\arguments{
\item{object}{ScEnsemble object with completed hypergraph lists}

\item{...}{Additional arguments}

\item{k}{Number of clusters (if NULL, will be estimated)}

\item{ensemble_methods}{Character vector of ensemble methods to apply}
}
\value{
ScEnsemble object with ensemble_results slot filled
}
\description{
Apply ensemble clustering methods
}
